Organize into Packages
- Organize all non-trivial classes into duke, exceptions, tasks packages
