+++ src/main/java/duke/Parser.java
@@ -1,3 +1,5 @@
+package duke;
+
 import exceptions.DukeException;
 import exceptions.InvalidDeadlineException;
 import exceptions.InvalidEventException;
