+++ src/main/java/duke/Ui.java
@@ -1,5 +1,7 @@
-import extensions.Task;
-import extensions.TaskList;
+package duke;
+
+import tasks.Task;
+import tasks.TaskList;
 
 public class Ui {
