+++ src/main/java/tasks/DeadlineTask.java
@@ -1,4 +1,4 @@
-package extensions;
+package tasks;
 
 import java.time.format.DateTimeFormatter;
 import java.time.LocalDateTime;
