+++ src/main/java/tasks/Task.java
@@ -1,4 +1,4 @@
-package extensions;
+package tasks;
 
 public abstract class Task {
     protected String desc;
